Move peerdb module to peerManager
